[Moneda] Se corrigieron los metodos agregar, actualizar y eliminar

+++ src/main/java/ar/com/draimo/jitws/service/MonedaService.java
@@ -2,46 +2,48 @@
 package ar.com.draimo.jitws.service;
 
 import ar.com.draimo.jitws.dao.IMonedaDAO;
+import ar.com.draimo.jitws.exception.MensajeRespuesta;
 import ar.com.draimo.jitws.model.Moneda;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 /**
  * Servicio Moneda
+ *
  * @author blas
  */
-
 @Service
 public class MonedaService {
-    
+
     //Define la referencia al dao
     @Autowired
     IMonedaDAO elementoDAO;
-    
+
     //Obtiene el siguiente id
     public int obtenerSiguienteId() {
         Moneda elemento = elementoDAO.findTopByOrderByIdDesc();
-        return elemento != null ? elemento.getId()+1 : 1;
+        return elemento != null ? elemento.getId() + 1 : 1;
     }
-    
+
     //Obtiene la lista completa
     public List<Moneda> listar() {
         return elementoDAO.findAll();
     }
-    
+
     //Obtiene una lista por nombre
     public List<Moneda> listarPorNombre(String nombre) {
-        return nombre.equals("***")? elementoDAO.findAll():
-            elementoDAO.findByNombreContaining(nombre);
+        return nombre.equals("***") ? elementoDAO.findAll()
+                : elementoDAO.findByNombreContaining(nombre);
     }
-    
+
     //Obtiene la moneda principal por defecto
     public Moneda obtenerPorDefecto() {
         return elementoDAO.findByPorDefectoTrue();
     }
-    
+
     //Establece la moneda como principal
     @Transactional(rollbackFor = Exception.class)
     public void establecerMonedaPrincipal(int idMoneda) {
@@ -49,41 +51,54 @@ public class MonedaService {
         moneda.setPorDefecto(false);
         elementoDAO.save(moneda);
     }
-    
+
     //Agrega un registro
     @Transactional(rollbackFor = Exception.class)
     public Moneda agregar(Moneda elemento) {
         elemento = formatearStrings(elemento);
-        if(elemento.getPorDefecto()) {
-            Moneda moneda = elementoDAO.findByPorDefectoTrue();
-            moneda.setPorDefecto(false);
-            elementoDAO.save(moneda);
+        Moneda moneda = elementoDAO.findByPorDefectoTrue();
+        if (moneda == null || moneda.getId() == elemento.getId()) {
+            elemento.setPorDefecto(true);
+        } else {
+            if (elemento.getPorDefecto()) {
+                moneda.setPorDefecto(false);
+                elementoDAO.save(moneda);
+            }
         }
-        return elementoDAO.saveAndFlush(elemento);
+        return elementoDAO.save(elemento);
     }
-    
+
     //Actualiza un registro
     @Transactional(rollbackFor = Exception.class)
     public void actualizar(Moneda elemento) {
         elemento = formatearStrings(elemento);
-        if(elemento.getPorDefecto()) {
-            Moneda moneda = elementoDAO.findByPorDefectoTrue();
-            moneda.setPorDefecto(false);
-            elementoDAO.save(moneda);
+        Moneda moneda = elementoDAO.findByPorDefectoTrue();
+        if (moneda == null || moneda.getId() == elemento.getId()) {
+            elemento.setPorDefecto(true);
+        } else {
+            if (elemento.getPorDefecto()) {
+                moneda.setPorDefecto(false);
+                elementoDAO.save(moneda);
+            }
         }
         elementoDAO.save(elemento);
     }
-    
+
     //Elimina un registro
     @Transactional(rollbackFor = Exception.class)
     public void eliminar(int elemento) {
-        elementoDAO.deleteById(elemento);
+        Moneda moneda = elementoDAO.findByPorDefectoTrue();
+        if (moneda.getId() == elemento) {
+            throw new DataIntegrityViolationException(MensajeRespuesta.ERROR_MONEDA_POR_DEFECTO);
+        } else {
+            elementoDAO.deleteById(elemento);
+        }
     }
-    
+
     //Formatea los strings
     private Moneda formatearStrings(Moneda elemento) {
         elemento.setNombre(elemento.getNombre().trim());
         return elemento;
     }
-    
+
 }
